refactor: move mocket.utils to mocket.core.utils
