+++ mocket/core/utils.py
